# code/basic-backend/src/app/middleware/swagger-output.json
{"swagger":"2.0","info":{"title":"My API","description":"API documentation","version":"1.0.0"},"host":"localhost:5000","basePath":"/","schemes":["http"],"paths":{"/auth/user/register":{"post":{"description":"","parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{"name":{"example":"any"},"email":{"example":"any"},"password":{"example":"any"},"role":{"example":"any"},"isActive":{"example":"any"},"RESET_PIN":{"example":"any"}}}}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/auth/user/login":{"post":{"description":"","parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{"name":{"example":"any"},"email":{"example":"any"},"password":{"example":"any"},"role":{"example":"any"},"isActive":{"example":"any"},"RESET_PIN":{"example":"any"}}}}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/auth/user/getAll":{"get":{"description":"","responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/auth/user/profile/{id}":{"get":{"description":"","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/auth/user/update/{id}":{"put":{"description":"","parameters":[{"name":"id","in":"path","required":true,"type":"string"},{"name":"body","in":"body","schema":{"type":"object","properties":{"password":{"example":"any"}}}}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/auth/user/delete/{id}":{"delete":{"description":"","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/auth/user/reset-password":{"post":{"description":"","parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{"name":{"example":"any"},"email":{"example":"any"},"password":{"example":"any"},"role":{"example":"any"},"isActive":{"example":"any"},"RESET_PIN":{"example":"any"}}}}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/auth/user/set-password":{"post":{"description":"","parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{"name":{"example":"any"},"email":{"example":"any"},"password":{"example":"any"},"role":{"example":"any"},"isActive":{"example":"any"},"RESET_PIN":{"example":"any"}}}}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/auth/user/forgot-password":{"post":{"description":"","parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{"name":{"example":"any"},"email":{"example":"any"},"password":{"example":"any"},"role":{"example":"any"},"isActive":{"example":"any"},"RESET_PIN":{"example":"any"}}}}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["user"]}},"/sales/product/create":{"post":{"description":"","parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{"name":{"example":"any"},"category":{"example":"any"},"images":{"example":"any"},"description":{"example":"any"},"price":{"example":"any"},"currency":{"example":"any"},"freeShipping":{"example":"any"},"addToCart":{"example":"any"},"isActive":{"example":"any"}}}}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["product"]}},"/sales/product/getAll":{"get":{"description":"","parameters":[{"name":"search","in":"query","type":"string"},{"name":"page","in":"query","type":"string"},{"name":"pageSize","in":"query","type":"string"},{"name":"column","in":"query","type":"string"},{"name":"direction","in":"query","type":"string"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["product"]}},"/sales/product/getAllHome":{"get":{"description":"","responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["product"]}},"/sales/product/getById/{id}":{"get":{"description":"","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["product"]}},"/sales/product/update/{id}":{"put":{"description":"","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["product"]}},"/sales/product/delete/{id}":{"delete":{"description":"","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"500":{"description":"Internal Server Error"}},"tags":["product"]}},"/sales/product/getAllMasterData":{"get":{"description":"","responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"}},"tags":["product"]}},"/sales/product/getAllProductsAddToCartCount":{"get":{"description":"","responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"}},"tags":["product"]}},"/sales/product/getAllProductsAddToCart":{"get":{"description":"","responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"}},"tags":["product"]}}}}